Fix ShellCheck complaint on Desktop Packages Check
The shellcheck complaint wants us to `grep -q` in line, however it also is complaining about the `-n` checks being done when it was an in line command call.

This change pulls the `apt | grep` into its own variable, then checks if that variable and its output is null/empty.  If it is, then it's not installed, otherwise throw the warning.

This should fix the CI failures as seen in other pull reqs on this line.